Guard against race in config listener test
These tests check whether the config subscription is working, but they
are prone to race conditions because subscribers only learn about config
changes asynchrnously. We put a 500ms sleep in place to give time for
the config update to propagate.

As an aside, we also modify the wait function so it only sends one
get_state message to the server and waits for the reply rather than
sending a repeated request every time it wakes up.